requirements.txt
Flask==2.3.3
pymongo==4.5.0
reportlab==4.0.4
python-dateutil==2.8.2
Werkzeug==2.3.7
google-generativeai==0.3.2
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-wtf==1.1.1
email_validator==2.3.0
